ai-coach/pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "duckduckgo-search>=7.2.1",
    "fastapi[standard]>=0.115.6",
    "google-generativeai>=0.8.3",
    "groq>=0.15.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "lancedb>=0.18.0",
    "ollama>=0.4.6",
    "openai>=1.59.8",
    "pgvector>=0.3.6",
    "phidata>=2.7.8",
    "psycopg[binary]>=3.2.4",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.13.0",
    "sentence-transformers>=3.4.0",
    "sqlalchemy>=2.0.37",
    "tantivy>=0.22.0",
    "uvicorn>=0.34.0",
    "watermark>=2.5.0",
]
